refactor(db): 将 bootcamp 数据相关文件移动到 lib/db 目录
- 将 data/bootcamp.ts 重命名为 lib/db/bootcamp.ts
- 更新相关文件中的导入路径，使用新的 lib/db/bootcamp 路径
- 优化课程标题，移除不必要的"第X周："前缀
